# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MedTime API",
    "description": "API do sistema MedTime para gestão de medicamentos e lembretes",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token de autenticação"
      }
    },
...
  },
  "paths": {
...
    "/medications": {
      "get": {
        "tags": [
          "medications"
        ],
        "description": "Lista todos os medicamentos do usuário",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "startDate": {
                        "type": "string"
                      },
                      "duration": {
                        "type": "number",
                        "nullable": true
                      },
                      "interval": {
                        "type": "number"
                      },
                      "isRecurring": {
                        "type": "boolean"
                      },
                      "totalQuantity": {
                        "type": "number"
                      },
                      "remainingQuantity": {
                        "type": "number"
                      },
                      "unit": {
                        "type": "string"
                      },
                      "dosageQuantity": {
                        "type": "number"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "reminders": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "scheduledFor": {
                              "type": "string"
                            },
                            "taken": {
                              "type": "boolean"
                            },
                            "takenAt": {
                              "type": "string",
                              "nullable": true
                            },
                            "skipped": {
                              "type": "boolean"
                            },
                            "skippedReason": {
                              "type": "string",
                              "nullable": true
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "scheduledFor",
                            "taken",
                            "takenAt",
                            "skipped",
                            "skippedReason",
                            "createdAt",
                            "updatedAt"
                          ],
                          "additionalProperties": false
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "description",
                      "startDate",
                      "duration",
                      "interval",
                      "isRecurring",
                      "totalQuantity",
                      "remainingQuantity",
                      "unit",
                      "dosageQuantity",
                      "userId",
                      "createdAt",
                      "updatedAt",
                      "reminders"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "medications"
        ],
        "description": "Cria um novo medicamento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "startTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "intervalPreset": {
                    "type": "string",
                    "enum": [
                      "6/6",
                      "8/8",
                      "12/12",
                      "24/24"
                    ]
                  },
                  "durationInDays": {
                    "type": "number",
                    "minimum": 0
                  },
                  "totalQuantity": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  },
                  "unit": {
                    "type": "string",
                    "enum": [
                      "comprimidos",
                      "ml",
                      "gotas",
                      "doses"
                    ]
                  },
                  "dosageQuantity": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 0
                  }
                },
                "required": [
                  "name",
                  "description",
                  "startTime",
                  "intervalPreset",
                  "durationInDays",
                  "totalQuantity",
                  "unit",
                  "dosageQuantity"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "number",
                      "nullable": true
                    },
                    "interval": {
                      "type": "number"
                    },
                    "isRecurring": {
                      "type": "boolean"
                    },
                    "totalQuantity": {
                      "type": "number"
                    },
                    "remainingQuantity": {
                      "type": "number"
                    },
                    "unit": {
                      "type": "string"
                    },
                    "dosageQuantity": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "reminders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "scheduledFor": {
                            "type": "string"
                          },
                          "taken": {
                            "type": "boolean"
                          },
                          "takenAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "skipped": {
                            "type": "boolean"
                          },
                          "skippedReason": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "scheduledFor",
                          "taken",
                          "takenAt",
                          "skipped",
                          "skippedReason",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "startDate",
                    "duration",
                    "interval",
                    "isRecurring",
                    "totalQuantity",
                    "remainingQuantity",
                    "unit",
                    "dosageQuantity",
                    "userId",
                    "createdAt",
                    "updatedAt",
                    "reminders"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/medications/{id}": {
      "get": {
        "tags": [
          "medications"
        ],
        "description": "Obtém uma medicação específica",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "startDate": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "number",
                      "nullable": true
                    },
                    "interval": {
                      "type": "number"
                    },
                    "isRecurring": {
                      "type": "boolean"
                    },
                    "totalQuantity": {
                      "type": "number"
                    },
                    "remainingQuantity": {
                      "type": "number"
                    },
                    "unit": {
                      "type": "string"
                    },
                    "dosageQuantity": {
                      "type": "number"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "reminders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "scheduledFor": {
                            "type": "string"
                          },
                          "taken": {
                            "type": "boolean"
                          },
                          "takenAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "skipped": {
                            "type": "boolean"
                          },
                          "skippedReason": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "scheduledFor",
                          "taken",
                          "takenAt",
                          "skipped",
                          "skippedReason",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "startDate",
                    "duration",
                    "interval",
                    "isRecurring",
                    "totalQuantity",
                    "remainingQuantity",
                    "unit",
                    "dosageQuantity",
                    "userId",
                    "createdAt",
                    "updatedAt",
                    "reminders"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "medications"
        ],
        "description": "Deleta uma medicação e seus lembretes",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}